Move left-click handling for object selection
It is the intention to have all input even handling on one place, that
is inside command.rs. Systems in this module send events to other
systems which implement low-level implementation.

With this slightly updated architecture, it will be easier to overload
left-click handling (e.g. selection vs construction of a new building).

Relates to #23.