Make it possible to keep a path target permanent
This will make it possible to set a destination and the entity will get
as close as possible after any change to map reach-ability.

Also, it is no longer possible to access current path target when new
path search is being computed / scheduled. This is a desired change
since the current path will be invalidated in few milliseconds anyways.

Relates to #24.