Upgrade to Bevy v0.10
This commit implements somewhat minimum number of changes necessary for
transition from Bevy v0.9 to Bevy v0.10. There are no intentional
changes to behavior except of the internal improvements made in Bevy
which constitute positive "side effects" of the upgrade (for example
improvements to rendering or parallel pipelined rendering).

We could drop dependency on iyes_loopless because it was superseded with
new scheduling in Bevy.

Glam was upgraded to match the version used by Bevy.

https://bevyengine.org/learn/book/migration-guides/0.9-0.10/

Relates to #324.