Fix attacking / laser cannon charging
Presence of Attacking component was used as an indicator that laser
cannon should be charged. However, Attacking component used to be
removed only when a) ChaseTargetComponent and b) the cannon did not had
clear sight of the target (e.g. when the target did no longer exist).
This did not work when ChaseTargetComponent was removed by another
system earlier.

Apart from that, fire ordering was incorrectly based on total
accumulated time instead of the time since one fire worth of charge was
reached.

This commit reworks and improves on laser charging. Apart from the fixes
to the unexpected behavior, it implements laser cannon discharging when
it is not actively charged.

Fixes #419.