Manufacturing: always stop at 100% completeness
When map capacity limit is reached, stop manufacturing only after full
completeness is reached. This is desirable so that the units are
delivered immediately after some of the existing units is destroyed.

It is the plan that the very same behavior will be used when spawn
location is occupied.

Relates to #422.