de_net: Discard Messages::port() method
This simplifies the code and goes with the fail-early principle.

Furthermore, as we are going to further break down Processor into
multiple independent tasks, we will need to log port from the other
tasks as well. These tasks won't need access to Messages for any other
reason so accepting port as an argument will make better sense for them.