de_net: Fix & improve resender termination (#511)
There was a bug with the original `break` -- it broke the for loop
instead the outer loop. This is now fixed with loop labels.

Apart from fixing the behavior, after this modification, resender keeps
running even if errors channel is closed as long as usender is still
running. Until this change, users were force to a) keep the channel to
avoid its closure and b) receive messages from it so it doesn't get
full. There are situations when the user is no longer interested in
connection errors but still needs the networking stack to keep running.
Now, it is possible to just drop the errors channel to get that effect.