Impl multiplayer shutdown on errors (#591)
Previously, game might have got into an inconsistent state or otherwise erroneous state (e.g. de-sync due to an unhandled network error). This is fixed by always shutting down multiplayer in such cases.

The implementation is a bit harsh as it shuts down the multiplayer on almost any error, including potentially recoverable issues. This might be improved in the future.