de_connector: Ignore message from non-participants (#592)
The main reason to ignore such messages is the case of a network issue
(from server to client). The affected player is automatically kicked
from the game in such a case. However, the client may not detect the
network issue (e.g. because there was no traffic from the client to the
server at the moment), therefore it may continue sending data to the
game. Such a state might lead to issues within other game participants
(e.g. receiving game sync messages from a unknown player).

Apart from the reason above, this is also a step towards better
security. However it won't suffice on its own.

Fixes #583.