de_connector: Limit clients to max one game (#601)
This fixes the issue of possible OpenGame package redelivery (which
would lead to multiple games being opened).

Also, it is a step towards protection against a DoS attack as one could
open infinite number of games and exhaust all ports as a result. To
solve this completely, it would be necessary to limit source IP
addresses not just sockets.

Relates to #547.