Improve shadows (#657)
This improves shadow quality because the cascade bounds are derived from
camera distance from the terrain. Most objects are close to the terrain
and the camera off nadir is limited, therefore majority of the shadows
are in a limited distance range dependent on the camera to focus point
distance.

Also, this change reduces number of cascades from 5 to 4, fixing this
warning: `The number of cascades configured for a directional light
exceeds the supported limit of 4.`

Fixes #574.