Multiplayer: Fix entity de-registering (#765)
This fixes: thread 'Compute Task Pool (4)' panicked at 'called
`Option::unwrap()` on a `None` value',
crates/multiplayer/src/playermsg.rs:270:78

Which was due to receiving a deregister message after the player left
the game. This can happen as a race-condition just after game end or if
the other player is kicked out due to network issues (but sends a
message around the time of the kick).

Fixes #757.