Upgrade to Bevy v0.13 (#790)
Migrated with respect to
https://bevyengine.org/learn/migration-guides/0-12-to-0-13/

Most notable changes are:

* All schedules (setup in `schedule.rs`) has been configured with
`auto_insert_apply_deferred: false` so that commands are applied only in
between individual schedules. This behavior is equal to the behavior
before the upgrade. This is required because the presence of components
of despawned entities is relied upon by e.g. `DespawnEventsPlugin`.

* Light brightness (setup in `map.rs`) was adjusted so it roughly
visually matches the lighting before the Bevy upgrade.

* Trail line transform (in `line.rs`) was updated since the base mesh is
now twice as big.

* Implementation of `DespawnEventsPlugin` and related events and systems
was changed because `Query::get_component::<T>()` is now deprecated. The
code was simplified and the despawn events are limited to a single
component only.