Add basic enemy colors to minimap (#795)
Fixes #584.

All players are assigned a color based on their player number. It also
changes the color of the units slightly to distinguish them from
buildings.

In the future I think it would be nice that each player could choose
their color from a list of curated colors that are visually distinctive,
and then they would always be consistent across single player or
multiplayer views. Or they are assigned a color, but then the UI would
need to change in other ways so they would know which color they were
assigned. This could be as simple as a color bar next to the their name
in the game list, or some other color indicator in the HUD.

I'm just learning Rust and Bevy, so hopefully I did things properly. Let
me know if you would like any changes.